Fixes
1.5.3

Fixes

1.5.4

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

"Rename"

OK

Update LICENSE

Address MacOS

MacOS

1.6.0

Update README.md

test script

1.6.1

Test macos works

1.6.2

Update README.md

Update FUNDING.yml

Update LICENSE

Update README.md

Update README.md

Update deps and correct setup script

Fixed some bugs and made more stable

Log

OK

ok

Upstream guys

Revert "Upstream guys"

This reverts commit 13d6aac3103902282dc40e9dbd9d537505aae615.

Changes from upstream

Updates

Relicense under AGPL

Good readme copy

Deps and scripts

ignore ups

Good

Security ups

IN total these commits close #38

Add body tag

Adding some image resize

Resize basics

Looks much better the viewing page

Download

OK

Some kind of issue with download

Good

great

Styles chipping forward

Ok styles are good!

Good

Debug script

ok

Logging events for debugging

ok

Update for debug

ok

Better

Bug with missing pdfs is fixed.

Trying new view styles

Even better try

Typo

Remove no need

ok

ok
