+++ .github/FUNDING.yml
@@ -1,3 +1,3 @@
 # These are supported funding model platforms
 
-github: crislin2046
+custom: https://dosyago.com/donate
