+++ public/index.html
@@ -1,6 +1,24 @@
 <!DOCTYPE html>
 <meta charset=utf-8>
 <meta name=viewport content=width=device-width,initial-scale=1>
-<title>BrowserGap SecureView&trade; Document to Image Server</title>
-<h1>BrowserGap SecureView&trade; Document to Image Server</h1>
+<title>DocumentSpark - BrowserBox SecureView&trade; Zero-Trust Document Viewing Service</title>
+<style>
+  :root {
+    font-family: system-ui, sans-serif;
+    background: linear-gradient(to bottom right, white, mintcream, mediumspringgreen, dodgerblue, navy);
+    background-size: 10% 10%;
+  }
+
+  .white-bg {
+    background: mistyrose;
+  }
+</style>
+<h1>DocumentSpark - <a href=https://github.com/dosyago/BrowserBoxPro>BrowserBox</a> SecureView&trade; Zero-Trust Document Viewing Service</h1>
+<p class=white-bg>
+  Find out more and buy a license: <a href=https://dosyago.com>DOSYAGO</a>
+</p>
+<p class=white-bg>
+  <a href=/secretpage-canneverbefound.html>Upload a document to view securely.</a>
+</p>
+
