+++ public/secretpage-canneverbefound.html
@@ -1,8 +1,9 @@
 <!DOCTYPE html>
+<html lang=en>
 <meta charset=utf-8>
 <meta name=viewport content=width=device-width,initial-scale=1>
 <meta name=theme-color content=#909>
-<title>PDF to Image Server</title>
+<title>DocumentSpark - BrowserBox SecureView&trade; Zero-Trust Document Viewing Service</title>
 <style>
   :root {
     xbackground: #000084;
@@ -57,51 +58,38 @@
     font-size: small;
   }
 </style>
-<form method=POST action=/very-secure-manifest-convert enctype=multipart/form-data>
-  <label>
-    <span class=hidden>Document</span>
-    <input required type=file name=pdf>
-    <input required type=hidden name=secret value=secret>
-  </label>
-  <button title="Upload file, get view link and begin conversion.">Convert</button>
+<h1>DocumentSpark - <a href=https://github.com/dosyago/BrowserBoxPro>BrowserBox</a> SecureView&trade; Zero-Trust Document Viewing Service</h1>
+<form id=uploader method=POST action=/very-secure-manifest-convert enctype=multipart/form-data target=output>
+  <fieldset>
+    <legend>Document Conversion for Viewing</legend>
+    <p>
+      <label>
+        File upload:
+        <input type=file name=pdf>
+      </label>
+      <strong>OR</strong>
+      <label>
+        URL for file:
+        <input type=url name=url placeholder="or a URL of a document">
+      </label>
+    <p>
+      <label>
+        Password/API Key for DocSpark Conversion Service:
+        <input required type=password name=secret value="Your password here">
+      </label>
+    <p>
+      <label>
+        Get link:
+        <button title="Upload file, get view link and begin conversion.">Begin Conversion</button>
+      </label>
+  </fieldset>
 </form>
-<p id=hello>
-Want to get new features? Get on the waitlist for the next version
-<a title="Get an invite" target=_blank href=https://forms.gle/YuQWniQrJmSVMoGF8>here.</a>
+<p>
+  Find out more and buy a license: <a href=https://dosyago.com>DOSYAGO</a>
+</p>
 <script>
-  switch(navigator.language.slice(0,2)) {
-    case "en": {
-      hello.innerHTML = `
-        Want to get new features? Get on the waitlist for the next version
-        <a title="Get an invite" target=_blank href=https://forms.gle/YuQWniQrJmSVMoGF8>here.</a>
-      `;
-      break;
-    }
-    case "ja": {
-      hello.innerHTML = `
-      他の機能が好きですか？ 少ない機能が必要ですか？<a href=https://secureview.cloudbrowser.xyz/uploads/filegr9k.56knk8r.jpg.html>LINE cris7fe</a>
-      `;
-      break;
-    }
-    case "zh": {
-      switch(navigator.language) {
-        case "zh-CN": {
-          hello.innerHTML = `
-          需要更多？ 文件寿命更长？ 更大的力量？ 在11月22日之前加入微信cris7fe，即可获得终身50％的折扣。
-          <a href=https://secureview.cloudbrowser.xyz/uploads/fileazjw.l7tq7e9.jpg.html>微信cris7fe</a>
-          `;
-          break;
-        }
-        case "zh-HK":
-        case "zh-TW": {
-          hello.innerHTML = `
-            需要更多？ 文件壽命更長？ 更大的力量？ 在11月22日之前加入微信cris7fe，即可獲得終身50％的折扣。
-            <a href=https://secureview.cloudbrowser.xyz/uploads/filegr9k.56knk8r.jpg.html>LINE cris7fe</a>
-          `;
-          break;
-        }
-      }
-    }
-  }
+  const secret = new URL(location).searchParams.get('secret');
+  const uploader = document.getElementById('uploader');
+  uploader.secret.value = secret;
 </script>
