+++ scripts/mac_install_deps.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+
+export HOMEBREW_NO_AUTO_UPDATE=1
+brew update
+brew install ghostscript 
+brew install mupdf-tools
+brew install imagemagick
+brew install --cask tex-live-utility
+brew install latexml
+brew install mactex
+brew install libreoffice
+
