+++ scripts/run-docspark.sh
@@ -1,0 +1,15 @@
+#!/bin/sh
+
+port=$1
+if [ -z $port ]; then
+  port=443
+  echo "Supply port, defaulting to 443"
+fi
+
+mkdir -p pdfs
+if [ ! -f "./pdfs/hashes.json" ]; then
+  echo "[]" > pdfs/hashes.json
+fi
+
+./public/uploads/clean.sh
+node src/index.js $port
