+++ secrets/README.md
@@ -1,3 +1,8 @@
 # Secrets
 
 key.js goes here
+
+it should look like
+
+
+module.exports = '<your key>';
